requirements.txt
pytest
pandas
openpyxl
appium-python-client
allure-pytest